requirements.txt
#python==3.7
torch
ray[tune]
ray[rllib]
gym
Pillow
pandas
dm-tree
keyboard
lz4
matplotlib
keyboard
pyyaml
scipy
